Added Data Storage template parmeter to layout classes
