+++ include/sparrow/layout/fixed_size_layout.hpp
@@ -21,8 +21,10 @@
 #include <utility>
 
 #include "sparrow/array/array_data.hpp"
+#include "sparrow/array/array_data_concepts.hpp"
 #include "sparrow/buffer/buffer.hpp"
 #include "sparrow/buffer/dynamic_bitset.hpp"
+#include "sparrow/layout/layout_iterator.hpp"
 #include "sparrow/utils/contracts.hpp"
 #include "sparrow/utils/iterator.hpp"
 #include "sparrow/utils/nullable.hpp"
