+++ test/test_array_data_concepts.cpp
@@ -18,13 +18,6 @@
 
 namespace sparrow
 {
-    static_assert(arrow_layout<fixed_size_layout<int>>);
-    static_assert(arrow_layout<variable_size_binary_layout<std::string, const std::string_view>>);
-    static_assert(arrow_layout<dictionary_encoded_layout<
-                      size_t,
-                      variable_size_binary_layout<std::string, const std::string_view>>>);
-    static_assert(!arrow_layout<std::string>);
-
     static_assert(range_of_arrow_base_type_extended<std::vector<int>>);
     static_assert(range_of_arrow_base_type_extended<std::vector<std::string>>);
     static_assert(range_of_arrow_base_type_extended<std::vector<std::string_view>>);
